@@ -2,6 +2,10 @@ package vyos
 
 import (
 	"context"
+	"encoding/json"
+	"reflect"
+	"sort"
+	"strings"
 	"time"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
@@ -35,15 +39,16 @@ func resourceConfigBlockTree() *schema.Resource {
 				ForceNew:         true,
 			},
 			"configs": {
-				Description: "Key/Value map of config parameters.",
+				Description: "Key/Value map of config parameters. Value can be a jsonencode list",
 				Type:        schema.TypeMap,
 				Elem: &schema.Schema{
 					Type: schema.TypeString,
 				},
 				Required:         true,
+				DiffSuppressFunc: configDiffSuppressFunc,
 			},
 		},
-        Timeouts: &schema.ResourceTimeout{
+		Timeouts: &schema.ResourceTimeout{
 			Create:  schema.DefaultTimeout(10 * time.Minute),
 			Read:    schema.DefaultTimeout(10 * time.Minute),
 			Update:  schema.DefaultTimeout(10 * time.Minute),
@@ -53,6 +58,57 @@ func resourceConfigBlockTree() *schema.Resource {
 	}
 }
 
+func configDiffSuppressFunc(k, old, new string, d *schema.ResourceData) bool {
+
+	multivalueOld := []string{}
+	err := json.Unmarshal([]byte(old), &multivalueOld)
+	if err != nil {
+		return false
+	}
+	sort.Strings(multivalueOld)
+
+	multivalueNew := []string{}
+	err = json.Unmarshal([]byte(new), &multivalueNew)
+	if err != nil {
+		return false
+	}
+	sort.Strings(multivalueNew)
+
+	return reflect.DeepEqual(multivalueOld, multivalueNew)
+}
+
+// Covert configs to a set of vyos client commands.
+// If expand_slice is set, then list values (json encoded) are expanded in multiple vyos client commands
+// If expand_slice is not set then the values in the map might contain slices
+func getCommandsForConfig(config interface{}, expand_slice bool) (commands map[string]any) {
+
+	commands = map[string]interface{}{}
+	for key, value := range config.(map[string]interface{}) {
+
+		// Try to decode the string as json list
+		value := value.(string)
+		multivalue := []string{}
+		err := json.Unmarshal([]byte(value), &multivalue)
+		if err == nil {
+			if expand_slice {
+				for _, subvalue := range multivalue {
+					commands[key+" "+subvalue] = ""
+				}
+			} else {
+				commands[key] = multivalue
+			}
+		} else {
+			// Could not decode json string - assume single value string
+			if expand_slice {
+				commands[key] = value
+			} else {
+				commands[key] = []string{value}
+			}
+		}
+	}
+	return
+}
+
 func resourceConfigBlockTreeCreate(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
 	var diags diag.Diagnostics
 
@@ -60,24 +116,10 @@ func resourceConfigBlockTreeCreate(ctx context.Context, d *schema.ResourceData, 
 	client := *p.client
 	path := d.Get("path").(string)
 
-	// Check if config already exists
-	configs, err := client.Config.ShowTree(ctx, path)
-	if err != nil {
-		return diag.FromErr(err)
-	}
-	
-	for attr, _ := range configs {
-		return diag.Errorf("Configuration block '%s' already exists and has '%s' set, try a resource import instead.", path, attr)
-	}
+	// Get commands needed to create resource in Vyos
+	commands := getCommandsForConfig(d.Get("configs"), true)
 
-	configs = d.Get("configs").(map[string]interface{})
-
-	commands := map[string]interface{}{}
-	for attr, val := range configs {
-		commands[path+" "+attr] = val
-	}
-
-	err = client.Config.SetTree(ctx, commands)
+	err := client.Config.Set(ctx, path, commands)
 	if err != nil {
 		return diag.FromErr(err)
 	}
@@ -91,10 +133,10 @@ func resourceConfigBlockTreeRead(ctx context.Context, d *schema.ResourceData, m 
 	var diags diag.Diagnostics
 
 	p := m.(*ProviderClass)
-	c := *p.client
+	//c := *p.client
 	path := d.Id()
 
-	configsTree, err := c.Config.ShowTree(ctx, path)
+	configsTree, err := p.ShowCached(ctx, path)
 	if err != nil {
 		return diag.FromErr(err)
 	}
@@ -104,13 +146,35 @@ func resourceConfigBlockTreeRead(ctx context.Context, d *schema.ResourceData, m 
 		return diag.FromErr(err)
 	}
 
-	configs := map[string]string{}
-	for _, path_value := range flat {
-		path := path_value[0]
-		value := path_value[1]
-		configs[path] = value
+	// Convert Vyos commands to Terraform schema
+	configs := map[string]interface{}{}
+	for _, config := range flat {
+		key := config[0]
+		value := config[1]
+		existing_value, ok := configs[key]
+		if ok {
+			// This is command with multiple values
+			switch existing_value := existing_value.(type) {
+			case string:
+				// Second value for command found - convert to slice
+				configs[key] = []string{existing_value, value}
+			case []string:
+				// N value for command found - append to slice
+				configs[key] = append(existing_value, value)
+			}
+		} else {
+			configs[key] = value
+		}
 	}
-	
+
+	// If there are slices then covert them to json strings
+	for key, value := range configs {
+		switch value := value.(type) {
+		case []string:
+			jsonBytes, _ := json.Marshal(value)
+			configs[key] = string(jsonBytes)
+		}
+	}
 
 	// Easiest way to allow ImportStatePassthroughContext to work is to set the path
 	if d.Get("path") == "" {
@@ -137,29 +201,103 @@ func resourceConfigBlockTreeUpdate(ctx context.Context, d *schema.ResourceData, 
 	old_configs := o.(map[string]interface{})
 	new_configs := n.(map[string]interface{})
 
-	deleted_attrs := []string{}
+	// Get commands needed to create old a new config
+	old_comands := getCommandsForConfig(old_configs, false)
+	new_comands := getCommandsForConfig(new_configs, false)
 
-	for old_attr := range old_configs {
-		value, ok := new_configs[old_attr]
-		_ = value
-		if !ok {
-			deleted_attrs = append(deleted_attrs, path+" "+old_attr)
+	// NOTE: it is important to apply new settings before deleting to
+	//       avoid errors. This is because delete and set are 2
+	//       different API calls and this might result in invalid
+	//       intermediary configs if we delete first.
+
+	// Calculate new commands (new config minus old config)
+	set_commands := map[string]interface{}{}
+	for command, new_value := range new_comands {
+		old_value, ok := old_comands[command]
+		if !ok || !reflect.DeepEqual(new_value, old_value) {
+			switch new_value := new_value.(type) {
+			case []string:
+				// List value - multiple subcommands
+				for _, value := range new_value {
+					set_commands[command+" "+value] = ""
+				}
+			case string:
+				set_commands[command] = new_value
+			}
+		}
+	}
+	if len(new_comands) > 0 {
+		errSet := c.Config.Set(ctx, path, new_comands)
+		if errSet != nil {
+			return diag.FromErr(errSet)
 		}
 	}
 
-	errDel := c.Config.Delete(ctx, deleted_attrs...)
-	if errDel != nil {
-		return diag.FromErr(errDel)
+	// Calculate delete commands (old config minus new config)
+	delete_commands := map[string]interface{}{}
+	for command, old_value := range old_comands {
+		new_value, ok := new_comands[command]
+		if !ok {
+			// Not found in new config - delete commpletly
+			if len(command) > len(path) {
+				// Do not delete path
+				delete_commands[command] = ""
+			}
+		} else {
+			// Compare old and new values for this command
+			for _, old_value_part := range old_value.([]string) {
+				found := false
+				for _, new_value_part := range new_value.([]string) {
+					if old_value_part == new_value_part {
+						found = true
+						break
+					}
+				}
+				// Only delete if old value not in new config AND this
+				// command is bellow the resource path
+				if !found && len(command) > len(path) {
+					delete_commands[command] = old_value_part
+				}
+			}
+		}
 	}
-
-	commands := map[string]interface{}{}
-	for attr, val := range new_configs {
-		commands[path+" "+attr] = val
+	// Remove orphan nodes as well
+	// An orphan is a node that does not have any entries in the new config
+	//
+	// Example: "system static-host-mapping host-name foo inet" => 1.2.3.4"
+	//          This will fail if we delete only "system static-host-mapping host-name foo inet"
+	//          and do not delete "system static-host-mapping host-name foo" as well
+	for key, _ := range delete_commands {
+		key_parts := strings.Split(key, " ")
+		parent := ""
+	out:
+		for _, key_part := range key_parts {
+			if len(parent) > 0 {
+				parent += " "
+			}
+			parent += key_part
+			found := false
+			for key, _ := range new_comands {
+				if strings.Contains(key, parent) {
+					// parent still exists in new config - keep parent
+					found = true
+					break
+				}
+			}
+			if !found && len(parent) > len(path) {
+				// Delete parent if not done already
+				if _, ok := delete_commands[parent]; !ok {
+					delete_commands[parent] = ""
+				}
+				break out
+			}
+		}
 	}
-
-	errSet := c.Config.SetTree(ctx, commands)
-	if errSet != nil {
-		return diag.FromErr(errSet)
+	if len(delete_commands) > 0 {
+		errDel := c.Config.Delete(ctx, path, delete_commands)
+		if errDel != nil {
+			return diag.FromErr(errDel)
+		}
 	}
 
 	p.conditionalSave(ctx)
