@@ -22,7 +22,7 @@ func dataSourceConfig() *schema.Resource {
 				Computed: true,
 			},
 		},
-        Timeouts: &schema.ResourceTimeout{
+		Timeouts: &schema.ResourceTimeout{
 			Read:    schema.DefaultTimeout(10 * time.Minute),
 			Default: schema.DefaultTimeout(10 * time.Minute),
 		},
@@ -31,18 +31,20 @@ func dataSourceConfig() *schema.Resource {
 
 func dataSourceConfigRead(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
 	p := m.(*ProviderClass)
-	c := *p.client
 	key := d.Get("key").(string)
 
-	value, err := c.Config.Show(ctx, key)
+	value, err := p.ShowCached(ctx, key)
 	if err != nil {
 		return diag.FromErr(err)
 	}
 
-	if value != nil {
-		if err := d.Set("value", *value); err != nil {
+	switch value := value.(type) {
+	case string:
+		if err := d.Set("value", value); err != nil {
 			return diag.FromErr(err)
 		}
+	default:
+		return diag.Errorf("Configuration at '%s' is not a string: %s.", key, value)
 	}
 
 	d.SetId(strconv.FormatInt(time.Now().Unix(), 10))
