@@ -3,7 +3,10 @@ package vyos
 import (
 	"context"
 	"crypto/tls"
+	"errors"
 	"net/http"
+	"strings"
+	"sync"
 	"time"
 
 	"github.com/foltik/vyos-client-go/client"
@@ -29,15 +32,21 @@ func Provider() *schema.Provider {
 				Optional: true,
 			},
 			"save": {
-				Type:     schema.TypeBool,
-				Optional: true,
-				Default:  true,
+				Type:        schema.TypeBool,
+				Optional:    true,
+				Default:     true,
 				Description: "Save after making changes in Vyos",
 			},
 			"save_file": {
-				Type:     schema.TypeString,
-				Optional: true,
+				Type:        schema.TypeString,
+				Optional:    true,
 				Description: "File to save configuration. Uses config.boot by default.",
+			},
+			"cache": {
+				Type:        schema.TypeBool,
+				Optional:    true,
+				Default:     true,
+				Description: "Use cache for read operations",
 			},
 		},
 		ResourcesMap: map[string]*schema.Resource{
@@ -53,9 +62,12 @@ func Provider() *schema.Provider {
 	}
 }
 
-type ProviderClass struct{
-	schema *schema.ResourceData;
-	client *client.Client;
+type ProviderClass struct {
+	schema *schema.ResourceData
+	client *client.Client
+
+	_showCacheMutex *sync.Mutex
+	_showCache      *map[string]any
 }
 
 func providerConfigure(ctx context.Context, d *schema.ResourceData) (interface{}, diag.Diagnostics) {
@@ -75,18 +87,53 @@ func providerConfigure(ctx context.Context, d *schema.ResourceData) (interface{}
 		c = client.NewWithClient(cc, url, key)
 	}
 
-	return &ProviderClass{d, c}, diag.Diagnostics{}
+	return &ProviderClass{d, c, &sync.Mutex{}, nil}, diag.Diagnostics{}
 }
 
 func (p *ProviderClass) conditionalSave(ctx context.Context) {
-	save      := p.schema.Get("save").(bool)
+	save := p.schema.Get("save").(bool)
 	save_file := p.schema.Get("save_file").(string)
 
-	if (save) {
+	if save {
 		if save_file == "" {
-			p.client.Config.Save(ctx);
+			p.client.Config.Save(ctx)
 		} else {
-			p.client.Config.SaveFile(ctx, save_file);
+			p.client.Config.SaveFile(ctx, save_file)
 		}
 	}
+}
+
+func (p *ProviderClass) ShowCached(ctx context.Context, path string) (any, error) {
+	cache := p.schema.Get("cache").(bool)
+
+	if !cache {
+		return p.client.Config.Show(ctx, path)
+	}
+
+	p._showCacheMutex.Lock()
+	if p._showCache == nil {
+		c := *p.client
+		showCache, err := c.Config.Show(ctx, "")
+		if err != nil {
+			return showCache, err
+		}
+		switch value := showCache.(type) {
+		case map[string]any:
+			p._showCache = &value
+		default:
+			return nil, errors.New("Configuration is not a map")
+		}
+	}
+	p._showCacheMutex.Unlock()
+
+	var val any = *p._showCache
+	for _, component := range strings.Split(path, " ") {
+		obj, ok := val.(map[string]any)[component]
+		if !ok {
+			return nil, nil
+		}
+		val = obj
+	}
+	return val, nil
+
 }
